non-completed vcpu

+++ rvm/src/arch/aarch64/regs.rs
@@ -1,7 +1,7 @@
 #[repr(C)]
 #[derive(Debug, Default, Clone)]
 pub struct GeneralRegisters {
-    pub x: [usize; 31],
+    pub x: [u64; 31],
 }
 
 macro_rules! save_regs_to_stack {
